backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.1
aioredis==2.0.1

# Elasticsearch
elasticsearch==8.11.0

# Celery for background tasks
celery==5.3.4
celery[redis]==5.3.4

# AI/ML
langchain==0.0.340
openai>=1.40.0
anthropic>=0.34.0
sentence-transformers==2.2.2
spacy==3.7.2

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
python-dateutil==2.8.2
pytz==2023.3

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Email
emails==0.6

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# News APIs
newsapi-python==0.2.6
praw==7.7.1
tweepy==4.14.0

# Google Trends
pytrends==4.9.2 